Manually merge skillController/Router & remove duplicated files, correct app.js endpoints
