+++ learn-link-api/app.js
@@ -6,7 +6,6 @@ const cors = require("cors");
 const { logger } = require("./middleware/logger");
 const { userRouter } = require("./routers/userRouter");
 const { appointmentRouter } = require("./routers/appointmentRouter")
-const { skillsRouter } = require('./routers/skillsRouter');
 const { skillRouter } = require("./routers/skillRouter");
 
 const app = express();
@@ -16,8 +15,7 @@ app.use(cors());
 app.use(logger);
 app.use("/api/users", userRouter);
 app.use("/api/appointments", appointmentRouter)
-app.use("/skills", skillsRouter); // Clash with "skillRouter" using /api/skills endpoint
-app.use("/skills", skillRouter);
+app.use("/skills", skillRouter);  // To-do: Change endpoint to "/api/skills/" (also change in dashboard-skills.js)
 
 app.get("/", (req, res) => {
   res.status(200).send("Welcome to Learn Link!");
