+++ learn-link-api/routers/skillRouter.js
@@ -10,6 +10,9 @@ skillRouter.post("/", skillController.createSkill);
 skillRouter.get("/:id", skillController.getSkillById);
 skillRouter.get("/", skillController.searchSkills);
 
+skillRouter.get("/current", skillController.currentSkillsInfo);
+skillRouter.get("/trending", skillController.topSkillsInfo);
+
 module.exports = {
   skillRouter,
 };
